+++ pysui/sui/sui_clients/sync_client.py
@@ -541,7 +541,12 @@ class SuiClient(_ClientMixin):
         return result
 
     def get_events(
-        self, *, query: SuiMap, cursor: str=None, limit: int=50, descending_order: bool=True
+        self,
+        *,
+        query: SuiMap,
+        cursor: str = "",
+        limit: int = 50,
+        descending_order: bool = True,
     ) -> SuiRpcResult:
         """get_events `sui_getEvents` API.
 
@@ -558,7 +563,7 @@ class SuiClient(_ClientMixin):
         """
         inargs: dict = locals().copy()
         inargs.pop("self")
-        return self.execute(QueryEvents(query=query, cursor=cursor, limit=limit, descending_order=descending_order, **inargs))
+        return self.execute(QueryEvents(**inargs))
 
     @deprecated(version="0.28.0", reason="Use SuiTransaction builder instead.")
     def pay_txn(